fix sort problem with sub sections

+++ backend/models/search/RecipeActiveRow.php
@@ -35,8 +35,8 @@ class RecipeActiveRow extends RecipeNutrientModel
         }
 
         usort($arr, static function ($a, $b) {
-            return ((int)$a->section->name > (int)$b->section->name)
-                || ($a->section->name === $b->section->name && ($a->catalog && $b->catalog && $a->catalog->id > $b->catalog->id));
+            return ((float)$a->section->name > (float)$b->section->name)
+                || ((float)$a->section->name === (float)$b->section->name && ($a->catalog && $b->catalog && $a->catalog->id > $b->catalog->id));
         });
         return $arr;
     }
